#4.9.0-rc3

+++ Cobilas.Unity/Graphics/IGU/Values/MouseButtonTypeEnum.cs
@@ -1,5 +1,5 @@
 ﻿namespace Cobilas.Unity.Graphics.IGU {
-    public enum MouseButtonType : byte {
+    public enum MouseButtonType : ushort {
         Left = 0,
         Right = 1,
         Mid = 2,
